Checkpoint before deparameterization of modules
This isn't a great place to checkpoint, but I'm about to rip the info
parameter out of every module. I still think it's the right proof
design, but it makes pretty much all the editor tools for Agda barf in
new and interesting ways every time I ask a question, so I'm temporarily
hardcoding it.